requirements.txt
absl-py==1.0.0
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
Bottleneck @ file:///C:/ci/bottleneck_1648010904582/work
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
Django @ file:///tmp/build/80754af9/django_1625585912945/work
easyocr==1.1.7
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
networkx==2.8
numexpr @ file:///C:/ci/numexpr_1640704337920/work
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373458095/work
Pillow==9.1.0
protobuf==3.20.1
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-bidi==0.4.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
tifffile==2022.5.4
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
